# openapi7.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Hamdaz ERP",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/teams/{ref}/access": {
   "get": {
    "tags": [
     "access"
    ],
    "summary": "What a team can reach",
    "operationId": "get_team_access_api_v1_teams__ref__access_get",
    "parameters": [
     {
      "name": "ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Ref"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TeamAccessOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "access"
    ],
    "summary": "Replace a team's whole module set",
    "operationId": "set_team_access_api_v1_teams__ref__access_put",
    "parameters": [
     {
      "name": "ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Ref"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SetAccess"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TeamAccessOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "access"
    ],
    "summary": "Grant one module to a team",
    "operationId": "grant_module_api_v1_teams__ref__access_post",
    "parameters": [
     {
      "name": "ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Ref"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/GrantModule"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TeamAccessOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/teams/{ref}/access/{module_key}": {
   "delete": {
    "tags": [
     "access"
    ],
    "summary": "Take a module away from a team",
    "operationId": "revoke_module_api_v1_teams__ref__access__module_key__delete",
    "parameters": [
     {
      "name": "ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Ref"
      }
     },
     {
      "name": "module_key",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Module Key"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/access/users/{user_id}": {
   "get": {
    "tags": [
     "access"
    ],
    "summary": "What someone else can reach",
    "operationId": "user_access_api_v1_access_users__user_id__get",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "User Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EffectiveAccessOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "EffectiveAccessOut": {
    "properties": {
     "user_id": {
      "type": "string",
      "format": "uuid",
      "title": "User Id"
     },
     "source": {
      "type": "string",
      "title": "Source"
     },
     "modules": {
      "items": {
       "$ref": "#/components/schemas/EffectiveModule"
      },
      "type": "array",
      "title": "Modules"
     },
     "via_teams": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Via Teams"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "source",
     "modules",
     "via_teams"
    ],
    "title": "EffectiveAccessOut"
   },
   "EffectiveModule": {
    "properties": {
     "key": {
      "type": "string",
      "title": "Key"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "admin_only": {
      "type": "boolean",
      "title": "Admin Only"
     },
     "pages": {
      "items": {
       "$ref": "#/components/schemas/PageOut"
      },
      "type": "array",
      "title": "Pages"
     }
    },
    "type": "object",
    "required": [
     "key",
     "name",
     "admin_only",
     "pages"
    ],
    "title": "EffectiveModule"
   },
...
   "GrantModule": {
    "properties": {
     "module_key": {
      "type": "string",
      "title": "Module Key"
     },
     "page_keys": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Page Keys"
     }
    },
    "type": "object",
    "required": [
     "module_key"
    ],
    "title": "GrantModule"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ModuleGrantOut": {
    "properties": {
     "module_key": {
      "type": "string",
      "title": "Module Key"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "all_pages": {
      "type": "boolean",
      "title": "All Pages"
     },
     "pages": {
      "items": {
       "$ref": "#/components/schemas/PageOut"
      },
      "type": "array",
      "title": "Pages"
     },
     "granted_at": {
      "type": "string",
      "format": "date-time",
      "title": "Granted At"
     },
     "granted_by_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Granted By Id"
     }
    },
    "type": "object",
    "required": [
     "module_key",
     "name",
     "all_pages",
     "pages",
     "granted_at",
     "granted_by_id"
    ],
    "title": "ModuleGrantOut"
   },
...
   "PageOut": {
    "properties": {
     "key": {
      "type": "string",
      "title": "Key"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "path": {
      "type": "string",
      "title": "Path"
     },
     "team_scoped": {
      "type": "boolean",
      "title": "Team Scoped",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "key",
     "name",
     "path"
    ],
    "title": "PageOut"
   },
...
   "SetAccess": {
    "properties": {
     "modules": {
      "additionalProperties": {
       "anyOf": [
        {
         "items": {
          "type": "string"
         },
         "type": "array"
        },
        {
         "type": "null"
        }
       ]
      },
      "type": "object",
      "title": "Modules"
     }
    },
    "type": "object",
    "title": "SetAccess",
    "description": "The team's complete access, replacing whatever it had."
   },
...
   "TeamAccessOut": {
    "properties": {
     "team_id": {
      "type": "string",
      "format": "uuid",
      "title": "Team Id"
     },
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "modules": {
      "items": {
       "$ref": "#/components/schemas/ModuleGrantOut"
      },
      "type": "array",
      "title": "Modules"
     }
    },
    "type": "object",
    "required": [
     "team_id",
     "slug",
     "name",
     "modules"
    ],
    "title": "TeamAccessOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}